django_app/requirements.txt
Django==2.2.9
django-haystack==2.8.1
rainbow_logging_handler==2.2.2
django-user-agents==0.4.0
git+https://github.com/lamby/django-email-from-template@master#egg=django_email_from_template
django-mathfilters==1.0.0
django_compressor==2.4
csscompressor==0.9.5
# slimit==0.8.1
django_inlinecss==0.3.0
# Required for openpgp encryption of keys
PGPy==0.4.3
# PGPy requires six, however, there is a bug: https://github.com/benjaminp/six/issues/210, so we had to downgrade to 1.10.0
# https://github.com/benjaminp/six/issues/210
# Have to use the binary, see https://stackoverflow.com/questions/48780354/python-django-error-version-glibc-private-not-defined
psycopg2-binary==2.8.4
six==1.14.0
Pillow==7.0.0
geopy==1.21.0
rethinkdb==2.4.4
boto==2.49.0
pyrax==1.9.8
apache-libcloud==2.8.1
django-ipware==2.1.0
elasticsearch>=2.0.0,<3.0.0
daemonocle==1.0.1
sox==1.3.7
gunicorn==20.0.4
sentry_sdk==0.14.2
python-memcached==1.59
docutils==0.16
python-dateutil==2.8.1
stripe==1.7.6   # old version
beautifulsoup4==4.8.2
click==7.0
# whitenoise==4.1.2
# django-cryptography==1.0
django-settings-export==1.2.1
gevent==1.4.0
django-cache-machine==1.1.0
django_newsletter==0.9
django-form-admin==0.5.1
django-sendgrid-v5==1.1.1
django-disposable-email-checker==2.0.2
quickemailverification==1.0.3
